Remove establishes end product (#6091)
