+++ app/controllers/intakes_controller.rb
@@ -38,6 +38,9 @@ class IntakesController < ApplicationController
   def complete
     intake.complete!(params)
     render json: intake.ui_hash
+
+    # TODO: This should probably be pushed into the model, since it is very
+    # end product specific
   rescue Caseflow::Error::DuplicateEp, Caseflow::Error::LongAddress => error
     render json: {
       error_code: error.error_code,
