+++ app/models/ama_review.rb
@@ -1,5 +1,4 @@
 class AmaReview < ApplicationRecord
-  include EstablishesEndProduct
   include CachedAttributes
 
   validate :validate_receipt_date
@@ -40,15 +39,6 @@ class AmaReview < ApplicationRecord
     return nil if contention_descriptions_to_create.empty?
     establish_end_product!
     create_contentions_on_new_end_product!
-  end
-
-  def end_product_description
-    end_product_establishment.description
-  end
-
-  def end_product_base_modifier
-    # This is for EPs not yet created or that failed to create
-    end_product_establishment.valid_modifiers.first
   end
 
   def veteran
