+++ app/models/higher_level_review.rb
@@ -6,7 +6,14 @@ class HigherLevelReview < AmaReview
 
   END_PRODUCT_MODIFIERS = %w[030 031 032 033 033 035 036 037 038 039].freeze
 
-  private
+  def end_product_description
+    end_product_establishment.description
+  end
+
+  def end_product_base_modifier
+    # This is for EPs not yet created or that failed to create
+    end_product_establishment.valid_modifiers.first
+  end
 
   def special_issues
     return [] unless same_office
@@ -22,11 +29,26 @@ class HigherLevelReview < AmaReview
     )
   end
 
-  def end_product_code
-    "030HLRR"
+  def establish_end_product!
+    end_product_establishment.perform!
+
+    update!(
+      end_product_reference_id: end_product_establishment.reference_id,
+      established_at: Time.zone.now
+    )
   end
 
-  def valid_modifiers
-    END_PRODUCT_MODIFIERS
+  private
+
+  def end_product_establishment
+    @end_product_establishment ||= EndProductEstablishment.new(
+      veteran: veteran,
+      reference_id: end_product_reference_id,
+      claim_date: receipt_date,
+      code: "030HLRR",
+      valid_modifiers: END_PRODUCT_MODIFIERS,
+      station: "397", # AMC
+      cached_status: end_product_status
+    )
   end
 end
