+++ spec/models/ramp_election_spec.rb
@@ -255,33 +255,6 @@ describe RampElection do
     end
   end
 
-  context "#established_end_product" do
-    subject { ramp_election.established_end_product }
-
-    let!(:other_ep) { Generators::EndProduct.build(veteran_file_number: veteran_file_number) }
-    let!(:matching_ep) { Generators::EndProduct.build(veteran_file_number: veteran_file_number) }
-
-    context "when matching end product has not yet been established" do
-      context "when end_product_reference_id is nil" do
-        it { is_expected.to be_nil }
-      end
-
-      context "when end_product_reference_id is set" do
-        let(:end_product_reference_id) { "not matching" }
-
-        it "raises EstablishedEndProductNotFound error" do
-          expect { subject }.to raise_error(RampElection::EstablishedEndProductNotFound)
-        end
-      end
-    end
-
-    context "when a matching end product has been established" do
-      let(:end_product_reference_id) { matching_ep.claim_id }
-
-      it { is_expected.to have_attributes(claim_id: matching_ep.claim_id) }
-    end
-  end
-
   context "#sync_ep_status!" do
     subject { ramp_election.sync_ep_status! }
